+++ dolphinscheduler-server/pom.xml
@@ -143,6 +143,11 @@
             <groupId>org.springframework</groupId>
             <artifactId>spring-test</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.apache.dolphinscheduler</groupId>
+            <artifactId>dolphinscheduler-register-api</artifactId>
+            <version>1.3.6-SNAPSHOT</version>
+        </dependency>
     </dependencies>
