+++ dolphinscheduler-spi/src/main/java/org/apache/dolphinscheduler/spi/plugin/DolphinPluginManagerConfig.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.dolphinscheduler.common.plugin;
+package org.apache.dolphinscheduler.spi.plugin;
 
 import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
